(improvement)(headless) Refactor MetricDrillDownChecker code
